Updated AAI to latest images - sparky
- sparky configuration files are separated from image,
and are in the aai/test-config project.
AAI and OOM teams agreed that minimal configuration
will be added to the image eventually, but until then all files are mapped.
- a change in logback.xml was made to match the logback for
aai/test-config project file for search-data, in order to get the pod up and
running with minimal changes to app files.

Issue-ID: OOM-947

Change-Id: Ife863ccd50615f40c21a9c72e666d4ab14c6011e
Signed-off-by: Keren Joseph <keren.joseph@amdocs.com>
Signed-off-by: Karen Joseph <keren.joseph@amdocs.com>
Signed-off-by: Keren Joseph <keren.joseph@amdocs.com>