Update patch set 3
Patch Set 3:

This is more of a portal question than a sparky question (as sparky simply redirects).

I see that the port 30225 is an open port (mapped to 8443) in the portal but I don't know the implications of doing the change so late.

The port 30215 (mapped to 8989) seems to still be open for http support as well.

If we want to go with https, it will need to be tested first.

Patch-set: 3