Update patch set 2
Patch Set 2: Code-Review-1

Hi, 

this should have been a draft to share the configuration and test it against CI builders. Unfortunately I could not add it as a draft, because the repository gave me "draft workflow is disabled" error message. That's why I gave it "-1" myself.

The pod cannot work because the jenkins docker build jobs for graphgraph are failing (due to python errors, I have no idea what it is and no way to investigate :( ..)

Patch-set: 2
Label: Code-Review=-1