Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> this fix seems a bit "overreacted" no?
> The initial goal was to have access to graph graph via node port so I think that https://gerrit.onap.org/r/c/aai/oom/+/95109 is a better way no?

haha, yes it might seem so :)

Going back and forth on the port issue in the end we decided to use the AAI proxy (which already is exposed via a nodeport) to route the traffic to graphgraph. This way we circumvent the problems with the nodeport and someone else can use it (as there are very few 302XX ports free and the AAI/OOM template forces to use a nodeport prefix although the nodeports are assigned to ONAP projects randomly i.e. multiple projects use a 302XX nodeport range)

Patch-set: 1