Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > regression on pods health in gating
> 
> Hi Morgan, 
> Please could you elaborate on this? I am not sure if the failure is due to my checkin.
> (note that this is still under review and verification though).
> Regards,
> Rajeev

that is why the gating is done...provide feedback before merge and avoid regressions when possible.
On your patch submission, we triggered teh deployment of a full ONAP and includes your patch instead of the corresponding master code.
We we do that it leads to issues several pods are not properly started and it seems related to your patch. (As with other patch there is no errors on pods/charts)

to get details on the errors we got, you can have a look at the traces as indicated in the gerrit review
all tests (A&AI healthcheck OK): https://orange-opensource.gitlab.io/lfn/onap/xtesting-onap-view/results/96352-1/index.html
pods check: https://orange-opensource.gitlab.io/lfn/onap/xtesting-onap-view/results/96352-1/k8s/onap-k8s.log
azt the end you got the logs and pods describe of the pods that did not start as expected

>>> Nb Failed Pods: 7
>>> List of Failed Pods: [onap-aai-aai-elasticsearch-c9f66db4b-zpbrq,onap-aai-aai-sparky-be-68bcf7c4c-p8k5c,onap-oof-oof-has-api-7b7f4c95b9-6fgvz,onap-oof-oof-has-controller-7db5bf5f8b-5bjhd,onap-oof-oof-has-data-8dc9c79dd-wpplv,onap-oof-oof-has-reservation-589cc655f4-v55qx,onap-oof-oof-has-solver-55d9849c69-s7dcq]

as your patch deals with these pods, I suspect that your change triggered a regression.
Error from server (BadRequest): container "aai-elasticsearch" in pod "onap-aai-aai-elasticsearch-c9f66db4b-zpbrq" is waiting to start: trying and failing to pull image

it is not 100% sure but as the patch before was clean, it has to be considered.

Please note that at the moment no gating chain is available. We got a hardware outage on our DC. We are working on a new chain on Azure but it is not ready yet.

Patch-set: 1