Use global storage templates for PVC
OOM has now templates in order to create the needed PVC, using:

* a PV with a specific class when using a common nfs mount path between
nodes (sames as today use) --> is the default behavior today
* or a storage class if we want to use dynamic PV.
  On this case, we use (in order of priority):
  - persistence.storageClassOverride if set on the chart
  - global.persistence.storageClass if set globally
  - persistence.storageClass if set on the chart

I've also created a PV/PVC for elasticsearch has it was hardcoded to
hostPath.

I've also removed hostPath for job logs (aai-traversal logs were not put
into the desired hostPath)

And I moved the 3 jobs of aai-graphadmin into 3 files as it's a better
practice (several object in the same file can lead to unexpected
behavior).

Behavior is supposed to be exactly the same as before.

Issue-ID: OOM-1227
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: I71c6cee8d7c33ef7bd39c40a43c26b377d4ceec3