Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2: Code-Review-1
> > > > 
> > > > As AAI is not part of "global" OOM repo, I need these informations in order to add it into the "oom_hardcoded_certificates.rst" file.
> > > > 
> > > > * Which components of AAI will use this AAI certificate? Is it AAI only?
> > > > * Do you also have "somewhere" onap-ca.crt? If yes, where please?
> > > > * Do you also have "somewhere" other certificates? If yes, where please?
> > > > 
> > > > When I have the answers, I'll add entries into the file, merge and we'll merge this one.
> > > 
> > > Are we adding new documentation requirements to the release after M4?  Is there a story for this?
> > 
> > For me it's part of REQ-215, part 4.10 (even if we're on the helm side).
> > So, we have two options:
> > * forbid hardcoded credentials now
> > * have a midway solution for Frankfurt only, what SECCOM has proposed.
> > 
> > It's up to you
> 
> it's also part of REQ-235 (no harcoded password)

OK.  I am currently in the middle of trying to figure out what all the certs in AAI are for, how to renew them, fix the ones that are broken, including how we use AAF.  Most of the devs who did the initial work on AAF and the non-core AAI resources (particularly on the fproxy/rproxy option) have left the project. So, this is a research project for me.  I probably won't have it before RC0.

Patch-set: 2