Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 3:
> 
> Healthdist is OK but when deploying basic VM (https://gating-results.onap.eu/results//105043-3/smoke-usecases/basic_vm/xtesting.debug.log) we're failing on early stage whereas "usually" we're failing later...
> 
> So I'll let @Morgan give its blessing (or not)

the situation is not ideal
to troubleshoot I need to test on a lab with the target pods (on a gating lab) - it means that just after a gating dealing with the patch, I must remove the lab from the pool to avoid than another gating overwritte the one I am interested to troubleshoot
or I need to set up a dedicated pod assuming that DT is also helping on the topic, it means 2 labs (in addition of our 2 daily Master labs so lots of resources)
 
so I think it will be easier to troubleshoot on the Daily Master. I suspect a change in datamodel at the beginning when retrieveing some infos from A&AI at the beginning of the tests (onap-tests is a bit stupid, if the API or the payload is changing, it will stop). 
It should not be a big issue.
It is too bad, we probably could have detected that earlier and then be able to fix it on the gating patch rather than merging on Master to troubleshoot on master
But we need more frequent release during the life cycle for than, which is not always easy.

Patch-set: 3
CC: Gerrit User 2679 <2679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+2