Update patch set 1
Patch Set 1:

This is a minor change of an expired certificate that is absolutely not going break anything and there is only upside to merging it because it's possible that we might make progress in getting portal integration with AAI working. But because the gating job is failing for reasons not related to this patch, both Portal and AAI remain blocked for **2 days**.  This is not acceptable.

Patch-set: 1
Reviewer: Gerrit User 101 <101@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 1803 <1803@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>