Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is a minor change of an expired certificate that is absolutely not going break anything and there is only upside to merging it because it's possible that we might make progress in getting portal integration with AAI working. But because the gating job is failing for reasons not related to this patch, both Portal and AAI remain blocked for **2 days**.  This is not acceptable.

The gating was failing on April 16th.
As the problem seemed to be an infrastructure infra issue, another gating was launched almost immediately and the results of the gating were OK on April 17th
you are allowed to merge when everything is fine all the more as the code is in aai/oom, which is under AAI responsability.
I think that OOM team is used to wait very frequently more than 2 days before getting feedback, and get new versions long time after the code freeze...
as it is difficult to evaluate for all the projects if the patchs are trivial or not, gating is the only sanity check procedure we have and infrastructure issues may occur...If you have any better alternatives feel free to share.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>