Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Hello James, I agree it’s not acceptable to wait for two days but:
> * I was off on Friday (and Tsc took the whole Thursday afternoon)
> * gating issue are more likely when a lot of patches gets uploaded. If AT&T is willing to offer some resources for gating, I’m sure we would have less pressure...
> * https://gerrit.onap.org/r/c/aai/oom/+/98933 took 4 days to get merged, so I totally understand your point. 
> 
> For your information, I plan to move back AAI into main OOM repository because:
> We need major changes on templates to make ONAP Kubernetes 1.16+ compatible, helm v3 compatible, storage efficient and service mesh compatible. 
> You’re the only repo (robot is already willing to come back to main repo) out of main repo and I believe most of our « kubernetes / helm experts » are on OOM project rather than on components projects. 
> 
> The difference will +1 from you (Or other AAI people when you submit) and then +2/merge for OOM committers.

The original direction was to move the charts out to the projects.  I thought the point of it was to allow committers on the projects to have flexibility in merging changes to their helm charts and remove the load from the OOM team. To me, this would be a good example of something that could be left to the project to decide - I did a component test in windriver and I was sure that it will not break the build, and it could be merged without risk. Instead, for this one trivial patch the job runs multiple times, chewing up unnecessary resources because I have tried to abide the requests of the OOM and Integration teams to allow the gating job to complete.

So, you've effectively stripped the autonomy away from the project committers and then unilaterally without discussion decide to completely pull control back into OOM. I wasn't clamoring for my own helm chart in the first place and pointing out that AAI is "the only repo out of the main repo" is not by my choice - I was asked by the OOM team to go first and I took it on because it seemed like the best for the project.

Patch-set: 1