Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > Hello James, I agree it’s not acceptable to wait for two days but:
> > * I was off on Friday (and Tsc took the whole Thursday afternoon)
> > * gating issue are more likely when a lot of patches gets uploaded. If AT&T is willing to offer some resources for gating, I’m sure we would have less pressure...
> > * https://gerrit.onap.org/r/c/aai/oom/+/98933 took 4 days to get merged, so I totally understand your point. 
> > 
> > For your information, I plan to move back AAI into main OOM repository because:
> > We need major changes on templates to make ONAP Kubernetes 1.16+ compatible, helm v3 compatible, storage efficient and service mesh compatible. 
> > You’re the only repo (robot is already willing to come back to main repo) out of main repo and I believe most of our « kubernetes / helm experts » are on OOM project rather than on components projects. 
> > 
> > The difference will +1 from you (Or other AAI people when you submit) and then +2/merge for OOM committers.
> 
> The original direction was to move the charts out to the projects.  I thought the point of it was to allow committers on the projects to have flexibility in merging changes to their helm charts and remove the load from the OOM team.

this was the plan yes. There are actually two types of changes related to helm charts:

* updating files (bumping version, certificates, although it should not be the case in Guilin, ...) which are usually done by the PTL.
  Then, we wait for gating and someone else (here the OOM comitters) will merge the change. It allows "other eyes" to look at it and merge

* give some consistency to all OOM charts. Here, the change are more complicated and requires a good knowledge of Helm but also it requires to know all the templates done in ONAP common folder to achieve some consistency.
  This is much complicated than the first type and often requires some adjustments between different folders of ONAP. Having them separated is a nightmare for us on "OOM main" as we may have to merge in parallel two commits (with the risks one break the other)
  It's particularly true with Robot (see CLI HC not working for days) but can happen also with AAI (I had an issue when working on Service Mesh where a change in AAI and SDC was required).

> To me, this would be a good example of something that could be left to the project to decide - I did a component test in windriver and I was sure that it will not break the build, and it could be merged without risk. Instead, for this one trivial patch the job runs multiple times, chewing up unnecessary resources because I have tried to abide the requests of the OOM and Integration teams to allow the gating job to complete.

Well, the I'm 100% sure it will work works 99% of a time but sometimes, we have made changes we forgot. The whole stuff of gating is to be sure that it still deploys. Yes it takes time, yes sometimes the deployment doesn't work because of gating system errors but it do give a good reliability on what OOM delivers.
Without gating, we would have a lot more errors. And doing it on "known to pass patches" allows also to have some "witness patches" to verify everything is correct.

> 
> So, you've effectively stripped the autonomy away from the project committers and then unilaterally without discussion decide to completely pull control back into OOM.
> I wasn't clamoring for my own helm chart in the first place and pointing out that AAI is "the only repo out of the main repo" is not by my choice - I was asked by the OOM team to go first and I took it on because it seemed like the best for the project.

I'm not blaming you, I'm just stating that you're the only one. Again, I do believe that having the whole OOM charts in the same repo is the best for the project.
We're talking of 23 patches for AAI since January, so roughly one every three days.
Out of these 23, 7 are for certificates, 6 are from OOM team.

So I'm not removing a lot of autonomy no?

The initial idea introduced in Dublin was to delegate to all the projects. 
In Frankfurt MSB tried it without lots of success (currently still facing Security and delivery issues). 

At the end, the drawbacks are more important than the advantages, mainly due to the difficulty to manage helm and get the full view on the packaging of common and other components. 
I do not see any interference in the project code, project are still fully master onboard of what they are delivering but if we want to be consistent, we must centralize the installation using the same rules/templates/best practices. 

As you did an effort to manage this part, I understand that it sounds strange to come back to the initial point, but from an installation perspective it makes sense. 
Most of the modular open source projects if not all do not distribute their installation in all the modules.

Patch-set: 1