Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Morgan, can you take a look at basic vm? It seems not to work but I’m not sure if this patch is the cause..;

On gating I suspect more an issue on var init in the gating chain.
Here the errors is on SDC call, no vendor is retrieved (A default vendor is supposed to be there) 
So I do not think it is due to the new version.

To be sure I can verify by updating the version on the daily Master where all the E2E tests seem to be stable now.

I do it that as soon as the Daily Master is finished and come back

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>