Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> e2e instantiation error in aai: https://gating-results.onap.eu/results//107283-1/smoke-usecases/basic_vm/xtesting.debug.log

hmm you are sure
https://gating-results.onap.eu/results//107283-1/smoke-usecases/basic_vm/xtesting.debug.log
looks good..

the test is PASS
maybe mislead by the trace level on A&AI call
when we ask the SO to create resources, we check in the A&AI that the resources are there
when we ask the SO to clean resources, we check in the A&AI that the resources are no more there

on resource creation, at the beginning we got 404 because the resource is not created yet (time between SO/A&A for the processing)..but after some time we got a 200 as expected 
at the end we got 200 and finally a 404 as expected (the resource has been cleaned)

the ERROR log level are misleading, it should not be ERROR..it is the expected behavior. That is why the test is PASS at the end.

Patch-set: 1
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>