Update patch set 1
Patch Set 1:

All right, don`t be too fast on this...
So this file here is targeted to define keys, and certain AAI artifacts following these keys. 
Now, the key values are used to discover different resource types in the TOSCA files created by SDC.

So first the SDC needs to avoid using openecomp within the TOSCA templates, which they are generating, and then the AAI "configuration files" can be adapted to work on these renewed models from SDC.

Ok, once we know this - are we good to merge this changes? Will?
As I have written before, we tested this internally, with SDC service models, which contain PNF resources. Works correctly. Actually, this merge request is blocking our submission in SO - SO-2640  ...

What would we need to merge this? Is there anything we could do to get this merged?

Patch-set: 1
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>