Prepare upgrade to readiness containers v3.0.1
Readiness 3.x is a new version, not using root user and using stable
Kubernetes API.
The upgrade is not doable by just changing version as script path has
moved.
Therefore, we must upgrade everything in one shot.
As AAI is not in the same repo, it's not doable for this component.
We then change variables only for AAI so we can continue to use older
images. We'll move AAI just after the rest has been upgraded.

Issue-ID: OOM-2545
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: I5afa83892043f4844afe12e61724a8d368a9f2e0