AAI-2958 Fix data-router in crash loopback state when there is no internet access
The problem is due to the fact in the data-router-oxm.xml file was referencing spring-context-4.2.xsd but it's not part of the spring-context-<xxx>.jar.  Therefore, it needs to reach out to the internet to download the internet.

And in spring 5, it started to use versionless spring-context.xsd and it's part of spring-context-5.1.13.jar (which is packaged in data-router 1.6.2).

The fix is to update data-router-oxm.xml to use the versionless spring-context.xsd (and it's only accessible via https).

Change-Id: Icf62978b73ae097fc3bcbaf3e96a8f63706dc4fc
Signed-off-by: Prudence Au <prudence.au@amdocs.com>
Issue-ID: AAI-2958