Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > @Sylvain, does the gating job also capture the log files somewhere?
> 
> That's exactly why we are pushing components to provide logs to stdout as k8s best practices says instead of some custom files 
> 
> > From the output of the basic_vm smoke test case, its not easy to determine the root cause of the issue.
> > If we can get access to log files on the system that are in filebeat side car pods /var/log/onap or if we can get more debugging info on what queries SDNC made to AAI and responses back from AAI, it would help debug this issue.

@Krzysztof, even if we change the logback, from the log output that we see so far, it doesn't seem to capture the logs at the end

https://logs.onap.org/onap-integration/gating//112257-1/infrastructure-healthcheck/k8s/pod-onap-aai-resources-5bf966d597-xdght-aai-resources-logs.html

From the link above, it only captured the log information up to 10:46 AM the smoke test ran from 11:20 AM to 11:28 AM, I don't see that time range being logged.
if there are standard output logs available for resources and traversal pod for that time range in the gating job output, can you provide that as that can help at least identify with the limited logging we have to the console?

Patch-set: 1