Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > @Dan Timoney seems to have fixed the SDNC issue on his side with a subsequent commit. Can we please look at https://gerrit.onap.org/r/c/aai/oom/+/112445 that addresses the SO issue? Reverting v20 and all our images back is going to cause heartache in Guilin for our functional use cases.
> 
> As far as I can see here:
> https://logs.onap.org/onap-integration/gating//112445-4/index.html
> 
> The review that you pointed us to still fails the basic_vm smoke test

BTW
We expect that the revert is only temporary. We want to revert to get green master once again and then you can introduce new images but this time making sure that gating is all the time green.

I've reverted everything till ~24.08:

https://gerrit.onap.org/r/c/oom/+/112484

Let's see if it works and then I'll be narrowing the scope of revert as much as possible

Patch-set: 1