Update patch set 3
Patch Set 3: Code-Review+1

On the 2 last gates, we just got an issue with DCAE deployment, we can see that 
- A&AI healthcheck tests are correct
- pnf-registrate is OK
- 5gbulkpm is OK
- basic_network is OK

the projects dealing with DCAE are failing
basic_onboard (introduced in H), basic_vm and basic_cnf are FAIL
but the gating is calling the master branch for xtesting-onap but is using the guilin xtesting dockers

in the guilin xtesting docker, basic_onboard is not included.

so some of the errors are due to a misalignement of the testing chain, the ansible roles are in master but the docker are in guilin. So far it was transparent but it is no more the case due to the facts that in H because of a bug in pbr/submodules we deploy the onaptests submodule using python setup.py developand no more pythonsetup.py install ..then all the overwritten path (to adapt the use case to the CI env) in the ansible roles are false...
we must come back to a standard module deployment

Conclusion for A&AI, I think it is good..no objection to merge as HC and robot based use cases that are using also AAI are fine.
It is good to keep 2 sources for smoke tests...

Patch-set: 3
Reviewer: Gerrit User 6079 <6079@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 2679 <2679@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1